Rewritten minecraft-common initscript
I was trying to fix the logfile-opening startup check code because it's broken,
and realised this is way too complicated. So the tmux wrapper is now gone and
replaced with a plain SSD invocation: if you need to poke the server directly,
use the built-in rcon interface. There's a nice perl5 module for it already.

I'd prefer to just do away with this and run via daemontools/runit, honestly.