Sorry, chpst -P is needed here after all.
runsvdir -P looked like it'd have the same effect, but no. It just calls setsid
on the runsv processes, agetty is greedy and wants setsid all for itself.

Not going to revert the /etc/runit/2 change, as it's still one less file to ship
with the ebuild and not hurting anything.