Add kexec reboot support to runit's shutdown stage
sysvinit has this functionality hard-coded in. This ought to be compatible with
the openrc kexec script (since that just loads the kernel and nothing else), but
you can also run /usr/sbin/kexec manually if preferred.

Should be safe if not used, as failing to kexec will just fall off the end of
the script and runit will reboot as normal.

(I had a specific use-case that needed this: my running kernel was missing NLS
modules necessary to mount /boot/efi, and couldn't insmod -f)