Yank out games.eclass from minecraft-server.ebuild
It's being too much of a pain in the ass. Just install the server as a normal
binary instead and let people deal with it themselves.

Requiring being in the "games" group to run this adds no security at all; it's
trivial to just grab the jar file off the website and run it manually.