sys-process/runit: Pre-emptively fix b.g.o 575478
This ebuild depends on far too many fragile and mismanaged Gentoo components.
At this point, the right thing to do seems to be to abandon ship entirely.
That probably means writing my own rc system to replace OpenRC. Should be easy.

On the subject of abandoning a sinking ship, this will be the last commit I push
to the github mirror of this repo.