runit: EAPI6; make shutdown more aggressive
I had a problem with automount hanging on exit, and didn't want to wait
5 minutes (or resort to sysrq) every time.

The correct way to solve that would require implementing a service
dependency tree like OpenRC which is far too much work for my limited
attention/motivation span. Instead I'm just going to make it 30 seconds.

This also throws in a "stty sane" because the console often isn't in a
usable state after X exits.