Makefile: don't run repoman in "make test"
I've had this as a git precommit hook for a while, and I **hate** it.

It's abysmally slow.
It complains about things irrelevant to overlays.
It stays silent on things that should be fixed.
It thinks it should own the workflow.

In short, it sucks. 0/10 would not recommend.

I'd put pkgcheck in here, but that doesn't return non-zero on errors,
even fatal ones, making it wholly unsuitable for automation.