audacious: remove old
In preparation for the impending 4.0 release, delete the old ebuilds and
supporting files. This implies removal of anything resembling GTK+2
support and the old autoconf-dependent plugins that still haven't been
ported to meson.

A new libarchive USE flag has been added, on by default. Read the flag
description for an explanation of why it's default.