media-fonts/noto-emoji-bin: drop 20200916.13.1
This fixes the pkgcheck gripe about the 13.1 version overshadowing the
older one. The only reason either of these are in here was so i could
grab them and put them on my phone, and the newer one isn't compatible
with it so there's no point keeping it around.

Also this stuff serves little purpose for a desktop any more because the
main noto-emoji package can now install the prebuilt font.