Resurrect gcolor2, purge GColor3
In very me style, i spent about an hour in the middle of the night
tweaking the contrib SVG icon in vim removing most of the inkscape crap
from it. It's about 75% smaller than the original (and keeps all the
metadata too).

Sick of GNOME software at this point. None of it fucking works.