Add sys-firmware/ath6kl-firmware
The upstream kernel has mostly the correct files for this but the driver
starts looping downwards from fw-5.bin, and kernel.org only supplies
fw-3.bin, so it spits a lot of unnecessary red text into dmesg.

Don't `rmmod` this driver or unplug the USB device, or the system will
freak out. In fact, don't plug it in or load the module at all.