sys-process/runit: backport changes to vanilla
In case anyone wants it, here's a -r14 that installs vanilla in the same
way the new ebuild installs void-runit.

Both versions also contain a fix to install etc/2 to examples/; this was
caught by QA messages here because vanilla symlinked to it from its
distro examples while void didn't.