sys-process/runit: fix GCC14 workaround
Merely sticking -Wno-* flags on the command line would only let it
compile, but /bin/sv would segfault consistently in certain codepaths
which is not a good thing to have happen at boot.

Setting -fpermissive seems to tweak GCC's semantics enough that it emits
binaries that don't crash. i don't understand how this works but it
works. Note that clang doesn't recognise this option, so it can't build
the code in its current state.