runit: various bug fixes and improvements
The previous attempt to migrate to /lib/runit/stages/ via a symlink
forest has major flaws so i'm removing it. The biggest problem was that
there's no good way to actually change the active stages at runtime
because a stage 1 may configure the system for a different stage 3 than
the one that ends up running. It's possible to deal with that by using
file bindmounts that shadow the new stages until next reboot, but at
that point it feels like i'm reinventing the worst parts of win32 for
negligible benefit.

This new design takes the existing scripts and breaks them into a
run-parts-like directory layout. A site operator that wants to replace
the openrc parts can then INSTALL_MASK or chmod -x them and drop in
suitable replacements, and those can be in any language.

It's boring and slightly ugly - on purpose.