www-plugins/passff-host-1.2.5: new ebuild
my half asleep ass unsuccessfully tried to file a b.g.o bug to get this
to support other browsers from an overlay and immediately got rebuked

So i did it myself, and updated it to a version that's not two years out
of date, and actually bothered to QA-lint the ebuild. The hardest part
of this process was getting pkgcheck to stop complaining about dosym.