sys-process/runit: revbump to fix GCC15 errors
…also to get the new stagescripts installed. "I have only proven this
code correct, not tried to run it", but i'll test it on myself before
anyone sees this so it should be fine.

Tried fixing the build failures properly, but it's easier and probably
safer to just slap an old -std on the CFLAGS and call it good.

Urge to rewrite this from scratch rising