sys-process/runit: the big refactor i wanted to do
The runit configuration is now closer to vanilla and supports use of
runsvchdir, should you have a need for it. The openrc configuration is
also closer to vanilla. You can mix and match both, they'll stay in
their own lanes and cgroups and won't step on each other.

Product has been thoroughly tested on animals