x11-base/wayback: new package
Compile and run tested (`wayback -d :1 -- enlightenment_start` from a
running X11 desktop) and it produced some output then segfaulted.

For a single 1kloc C file, that's not terrible.

(i am well aware of Ariadne's long, long history of throwing shit over
the wall and then abandoning it unfinished a month later as soon as the
next thing to edgily subpost about comes along. i have no reason to
believe this time won't turn out exactly the same. But this took zero
effort to ebuild and it'll take zero effort to rm later, and if it gives
fash trash a black eye in the process then all the better.)